Move methods to services
