+++ lib/src/services/library_items_service.dart
@@ -1,5 +1,34 @@
+import 'dart:convert';
+
+import 'package:http/http.dart' as http;
+
+import '../models/abs_audiobook.dart';
+import '../models/abs_play_item_request.dart';
 import 'service.dart';
 
 class LibraryItemsService extends Service {
   const LibraryItemsService(super.api);
+
+  Future<AbsAudiobook> getLibraryItem(String libraryItemId) async {
+    http.Response response = await api.client.get(
+      api.createUri(
+        api.baseUrl,
+        '/api/items/$libraryItemId',
+        {'expanded': '1'},
+      ),
+      headers: api.authJsonHeader,
+    );
+
+    return AbsAudiobook.fromJson(jsonDecode(utf8.decode(response.bodyBytes)));
+  }
+
+  Future<String> playLibraryItem(
+      String id, AbsPlayItemRequest playRequest) async {
+    http.Response response = await api.client.post(
+      api.createUri(api.baseUrl, '/api/items/$id/play'),
+      headers: api.authJsonHeader,
+      body: jsonEncode(playRequest.toJson()),
+    );
+    return jsonDecode(response.body)['id'];
+  }
 }
