+++ lib/src/services/me_service.dart
@@ -1,5 +1,37 @@
+import 'dart:convert';
+
+import 'package:http/http.dart' as http;
+
+import '../abs_audiobook_progress.dart';
 import 'service.dart';
 
 class MeService extends Service {
   const MeService(super.api);
+
+  Future markPlayed(String itemId) async {
+    (await patchAudiobook(itemId, true));
+  }
+
+  Future markUnplayed(String itemId) async {
+    await patchAudiobook(itemId, false);
+  }
+
+  Future<http.Response> patchAudiobook(String itemId, bool isRead) async {
+    return await api.client.patch(
+      api.createUri(
+        api.baseUrl,
+        '/api/me/progress/$itemId',
+      ),
+      headers: api.authJsonHeader,
+      body: utf8.encode(jsonEncode({'isFinished': isRead})),
+    );
+  }
+
+  Future updateProgress(AbsAudiobookProgress progress) async {
+    await api.client.patch(
+      api.createUri(api.baseUrl, '/api/me/progress/${progress.id}'),
+      headers: api.authJsonHeader,
+      body: utf8.encode(jsonEncode(progress.toJson())),
+    );
+  }
 }
